Moved Unavailable evidence to avoid circular import issues
